requirements.txt
fastapi == 0.95.1
spotipy == 2.23.0
billboard-py == 7.0.0
uvicorn == 0.22.0
databases==0.5.2
sqlalchemy==1.4.26
python-dotenv==0.19.1
asyncpg==0.27.0
psycopg2-binary~=2.9.3
cryptography==38.0.0
